requirements.txt
httpx>=0.27.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0
aiohttp>=3.9
